[1552]: Move the mockNameKeeper to its own file since it's not genesis specific anymore and add a missing function.
